incomplete with bugs

+++ serial/util.h
@@ -11,8 +11,8 @@ using namespace std;
 
 #define PI 3.14159265
 
-vector<int> extract_int(string str, string delimiter);
-vector<vector<int> > read_inputs();
-vector<int> read_weights_and_bias();
+vector<double> extract_int(string str, string delimiter);
+vector<vector<double> > read_inputs();
+vector<double> read_weights_and_bias();
 
 #endif
